weihnachtsaufgabe-requirements.txt
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
pytesseract==0.3.13
